# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Auto Anime Downloader daemon",
        "title": "Auto Anime Downloader API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8091",
    "basePath": "/api/v1",
    "paths": {
...
        "/config": {
            "get": {
                "description": "Returns the current daemon configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/files.Config"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the daemon configuration with the provided values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Update configuration",
                "parameters": [
                    {
                        "description": "Configuration object",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/files.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
...
        "/daemon/start": {
            "post": {
                "description": "Starts the daemon loop for automatic anime checking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daemon"
                ],
                "summary": "Start daemon",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/daemon/stop": {
            "post": {
                "description": "Stops the daemon loop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daemon"
                ],
                "summary": "Stop daemon",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "files.Config": {
            "type": "object",
            "properties": {
                "anilist_username": {
                    "type": "string"
                },
                "anilist_usernames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "anime_ids_are_media_ids": {
                    "description": "AnimeIDsAreMediaIDs marca que daemon.MigrateAnimeIDsToMedia ja converteu os AnimeID\ngravados de id de ENTRADA (MediaList, por conta) para id de MIDIA (ver decisions.md #43).\nO default e false de proposito: um config.json anterior a este campo desserializa por\ncima do default e precisa migrar. Numa instalacao nova a migracao roda sem nada a fazer\ne liga o campo no primeiro passe.",
                    "type": "boolean"
                },
                "check_interval": {
                    "type": "integer"
                },
                "completed_anime_path": {
                    "type": "string"
                },
                "delete_statuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "delete_watched_episodes": {
                    "type": "boolean"
                },
                "download_media_statuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "download_statuses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "episode_retry_limit": {
                    "type": "integer"
                },
                "excluded_list": {
                    "type": "string"
                },
                "excluded_lists": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_batch_episodes": {
                    "description": "MaxBatchEpisodes e o teto de episodios para usar batch. Acima disso um anime finalizado\nvolta ao caminho episodio-a-episodio (onde MaxEpisodesPerAnime vale). 0 desliga o teto.",
                    "type": "integer"
                },
                "max_batch_torrent_size_gb": {
                    "description": "MaxBatchTorrentSizeGB / MaxEpisodeTorrentSizeGB descartam da busca torrents acima do teto,\nem GiB. 0 desliga (default: nao filtrar, pois nao existe numero defensavel sem saber se o\nusuario quer 1080p web ou remux).",
                    "type": "number"
                },
                "max_concurrent_downloads": {
                    "description": "MaxConcurrentDownloads limita quantos torrents INCOMPLETOS baixam ao mesmo tempo; o\nexcedente espera na fila (torrents.queue). 0 desliga o limite. Seeding nunca conta.\n\nNao precisa de migracao: LoadConfigs desserializa POR CIMA de getDefaultConfig(),\nentao um config.json anterior a este campo ja carrega valendo o default.",
                    "type": "integer"
                },
                "max_episode_torrent_size_gb": {
                    "type": "number"
                },
                "max_episodes_per_anime": {
                    "description": "MaxEpisodesPerAnime limita quantos episodios de um anime existem ao mesmo tempo, e vale\nAPENAS no caminho episodio-a-episodio: um batch e um torrent so, entao limitar registros\nnao limitaria bytes nem arquivos na biblioteca (ver decisions.md).",
                    "type": "integer"
                },
                "min_free_disk_percent": {
                    "description": "MinFreeDiskPercent barra a adicao de novos torrents abaixo dessa porcentagem de espaco\nlivre no volume da biblioteca. 0 desliga.",
                    "type": "integer"
                },
                "notifications": {
                    "$ref": "#/definitions/files.NotificationsConfig"
                },
                "priorities": {
                    "$ref": "#/definitions/nyaa.Priorities"
                },
                "rename_files_for_jellyfin": {
                    "type": "boolean"
                },
                "watched_episodes_to_keep": {
                    "type": "integer"
                }
            }
        },
        "files.NotificationsConfig": {
            "type": "object",
            "properties": {
                "batch_window_seconds": {
                    "description": "BatchWindowSeconds agrupa os eventos de uma mesma janela num webhook so. 0 desliga o\nagrupamento (um webhook por evento, comportamento original). Existe porque um backfill\nde biblioteca dispara um `download_completed` por torrent e estoura a cota de servicos\ncomo o ntfy.sh — ver decisions.md #47.",
                    "type": "integer"
                },
                "webhooks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/files.WebhookPreset"
                    }
                }
            }
        },
        "files.WebhookPreset": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "nyaa.Priorities": {
            "type": "object",
            "properties": {
                "audio": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "codecs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "criteria_order": {
                    "description": "CriteriaOrder é a ordem em que os critérios de sort são aplicados.\nValores válidos: uncensored, source, resolution, codec, fansub, audio, health, size.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fansubs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ignore_list": {
                    "description": "IgnoreList: substrings (case-insensitive) que descartam o release de vez.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}